add lazy data resolver
