+++ pyproject.toml
@@ -167,8 +167,6 @@ opentelemetry-exporter-otlp = ">=1.25.0"
 opentelemetry-sdk = ">=1.25.0"
 uvicorn = ">=0.24.0"
 uvloop = ">=0.17.0"
-
-[tool.poetry.group.transfer.dependencies]
 truss_transfer="0.0.1rc3"
 
 [build-system]
