+++ truss/tests/templates/core/server/test_lazy_data_resolver_v2.py
@@ -4,7 +4,7 @@ import time
 from pathlib import Path
 
 import pytest
-from truss.templates.shared.lazy_data_resolver_v2 import (
+from truss.templates.shared.lazy_data_resolver import (
     TRUSS_TRANSFER_AVAILABLE,
     LazyDataResolverV2,
 )
